+++ src/components/TypeaheadSelect/TypeaheadSelect.tsx
@@ -1,4 +1,4 @@
-import { Combobox } from '@headlessui/react';
+import { Combobox, Transition } from '@headlessui/react';
 import classNames from 'classnames';
 import isEqual from 'lodash.isequal';
 import { Fragment, KeyboardEvent, RefObject, useEffect, useRef } from 'react';
@@ -6,7 +6,7 @@ import { FieldValues } from 'react-hook-form';
 import { Input, Progress } from 'react-daisyui';
 import { useTypeaheadSelect } from './useTypeaheadSelect';
 import { Badge } from '../Badge';
-import { Dropdown, DropdownMenu, DropdownOption } from '../Dropdown';
+import { DropdownMenuItem } from '../DropdownMenu';
 import { FormError, FormFieldProps, FormLabel } from '../Form';
 import { ComboboxMulti } from '../Form/FormComboboxMulti';
 import { ComboboxSingle } from '../Form/FormComboboxSingle';
@@ -25,6 +25,7 @@ export interface TypeaheadSelectProps<
   getItemValue?: (data: DataItem) => string;
   inputPlaceholder?: string;
   inputRef?: RefObject<HTMLInputElement>;
+  menuClassName?: string;
   multi?: true;
 }
 
@@ -44,6 +45,7 @@ export const TypeaheadSelect = <
   inputRef,
   isRequired,
   labelText,
+  menuClassName,
   multi,
   name,
   onDebouncedChange,
@@ -64,7 +66,7 @@ export const TypeaheadSelect = <
     setShowDropdown,
     setSelectedItems,
     setQuery,
-  } = useTypeaheadSelect<DataItem, Values>({
+  } = useTypeaheadSelect<HTMLUListElement, DataItem, Values>({
     controllerProps,
     debounceTime,
     name,
@@ -86,7 +88,7 @@ export const TypeaheadSelect = <
   }, [defaultOptions]);
   return (
     <Component
-      className={className}
+      className={classNames('relative', className)}
       getItemValue={getItemValue}
       name={name}
       selectedItems={selectedItems}
@@ -96,113 +98,120 @@ export const TypeaheadSelect = <
         if (multi === undefined) ref.current?.focus();
       }}
     >
-      {labelText !== undefined ? (
-        <Combobox.Label as={FormLabel} isRequired={isRequired}>
-          {labelText}
-        </Combobox.Label>
-      ) : null}
-      {enableBadges ? (
-        <div
-          className={classNames(
-            'input-bordered input flex cursor-pointer items-center gap-1 overflow-x-scroll scrollbar-none',
-            `input-${fieldColor ?? 'ghost'}`,
-          )}
-          onBlur={event => {
-            if (event.relatedTarget === null) setShowDropdown(false);
-          }}
-          onClick={() => setShowDropdown(true)}
-          onKeyDown={event => {
-            if (event.key === 'Enter') {
-              event.preventDefault();
-              setShowDropdown(true);
-            } else if (event.key !== 'Tab' && event.key !== 'Shift') {
-              setShowDropdown(true);
-            }
-          }}
-          ref={ref}
-          tabIndex={0}
-        >
-          {selectedItems.length > 0
-            ? selectedItems.map((item, index) => (
-                <Badge
-                  dismissable
-                  key={item.id}
-                  onDismiss={() => clearSelectedItem(index)}
-                >
-                  {getBadgeText?.(item) ?? getItemText(item)}
-                </Badge>
-              ))
-            : placeholder}
-        </div>
-      ) : (
-        <Combobox.Input
-          as={Input}
-          className="w-full"
-          onChange={({ target: { value } }) => {
-            setQuery(value);
-            if (value.length > 0) {
-              setShowDropdown(true);
-            } else {
-              setShowDropdown(false);
-              setSelectedItems([]);
-            }
-          }}
-          placeholder={placeholder}
-          ref={ref}
-        />
-      )}
-      <Dropdown
-        className="text-left"
-        onKeyDown={({ key }) => {
-          if (key === 'Escape') setShowDropdown(false);
-        }}
-        ref={dropdownRef}
-      >
-        {showDropdown ? (
-          <Combobox.Options
-            as={DropdownMenu}
-            className="bg-base-100 shadow-xl"
-            static
-          >
-            {enableBadges ? (
-              <Combobox.Input
-                as={Input}
-                className="w-full"
-                onChange={({ target: { value } }) => setQuery(value)}
-                onKeyDown={({ key }: KeyboardEvent) => {
-                  if (key === 'Tab') setShowDropdown(false);
-                }}
-                placeholder={inputPlaceholder}
-                ref={searchInputRef}
-                value={query}
-              />
-            ) : null}
-            {isLoading ? (
-              <Progress className={classNames(enableBadges && 'mt-2')} />
-            ) : null}
-            {!isLoading && options?.length === 0 ? (
-              <DropdownOption disabled>No Results</DropdownOption>
-            ) : null}
-            {!isLoading &&
-              options?.map((option, index) => (
-                <Combobox.Option as={Fragment} key={option.id} value={option}>
-                  {({ active, disabled, selected }) => (
-                    <DropdownOption
-                      active={active}
-                      className={classNames(
-                        index === 0 && enableBadges && 'mt-2',
-                      )}
-                      disabled={disabled}
-                      selected={multi === true ? selected : undefined}
-                    >
-                      {getItemText(option)}
-                    </DropdownOption>
-                  )}
-                </Combobox.Option>
-              ))}
-          </Combobox.Options>
+      <div className="form-control">
+        {labelText !== undefined ? (
+          <Combobox.Label as={FormLabel} isRequired={isRequired}>
+            {labelText}
+          </Combobox.Label>
         ) : null}
-      </Dropdown>
+        {enableBadges ? (
+          <div
+            className={classNames(
+              'input-bordered input flex cursor-pointer items-center gap-1 overflow-x-scroll scrollbar-none',
+              `input-${fieldColor ?? 'ghost'}`,
+            )}
+            onBlur={event => {
+              if (event.relatedTarget === null) setShowDropdown(false);
+            }}
+            onClick={() => setShowDropdown(true)}
+            onKeyDown={event => {
+              if (event.key === 'Enter') {
+                event.preventDefault();
+                setShowDropdown(true);
+              } else if (event.key !== 'Tab' && event.key !== 'Shift') {
+                setShowDropdown(true);
+              }
+            }}
+            ref={ref}
+            tabIndex={0}
+          >
+            {selectedItems.length > 0
+              ? selectedItems.map((item, index) => (
+                  <Badge
+                    dismissable
+                    key={item.id}
+                    onDismiss={() => clearSelectedItem(index)}
+                  >
+                    {getBadgeText?.(item) ?? getItemText(item)}
+                  </Badge>
+                ))
+              : placeholder}
+          </div>
+        ) : (
+          <Combobox.Input
+            as={Input}
+            className="w-full"
+            onChange={({ target: { value } }) => {
+              setQuery(value);
+              if (value.length > 0) {
+                setShowDropdown(true);
+              } else {
+                setShowDropdown(false);
+                setSelectedItems([]);
+              }
+            }}
+            placeholder={placeholder}
+            ref={ref}
+          />
+        )}
+      </div>
+      <Transition
+        as={Fragment}
+        show={showDropdown}
+        enter="transition duration-100 ease-out"
+        enterFrom="transform scale-95 opacity-0"
+        enterTo="transform scale-100 opacity-100"
+        leave="transition duration-75 ease-out"
+        leaveFrom="transform scale-100 opacity-100"
+        leaveTo="transform scale-95 opacity-0"
+      >
+        <Combobox.Options
+          as="ul"
+          className={classNames(
+            'menu rounded-box absolute z-50 bg-base-100 p-2 shadow-xl',
+            menuClassName,
+          )}
+          ref={dropdownRef}
+          static
+        >
+          {enableBadges ? (
+            <Combobox.Input
+              as={Input}
+              className="w-full"
+              onChange={({ target: { value } }) => setQuery(value)}
+              onKeyDown={({ key }: KeyboardEvent) => {
+                if (key === 'Tab') setShowDropdown(false);
+              }}
+              placeholder={inputPlaceholder}
+              ref={searchInputRef}
+              value={query}
+            />
+          ) : null}
+          {isLoading ? (
+            <Progress className={classNames(enableBadges && 'mt-2')} />
+          ) : null}
+          {!isLoading && options?.length === 0 ? (
+            <DropdownMenuItem disabled>No Results</DropdownMenuItem>
+          ) : null}
+          {!isLoading &&
+            options?.map((option, index) => (
+              <Combobox.Option as={Fragment} key={option.id} value={option}>
+                {({ active, disabled, selected }) => (
+                  <DropdownMenuItem
+                    active={active}
+                    className={classNames(
+                      index === 0 && enableBadges && 'mt-2',
+                    )}
+                    disabled={disabled}
+                    selected={multi === true ? selected : undefined}
+                  >
+                    {getItemText(option)}
+                  </DropdownMenuItem>
+                )}
+              </Combobox.Option>
+            ))}
+        </Combobox.Options>
+      </Transition>
       {error?.message !== undefined ? (
         <FormError>{error.message}</FormError>
       ) : null}