+++ src/components/index.ts
@@ -1,6 +1,6 @@
 export * from './AlertMessages';
 export * from './Badge';
-export * from './Dropdown';
+export * from './DropdownMenu';
 export * from './Form';
 export * from './FullScreenLoader';
 export * from './Icons';
@@ -8,5 +8,7 @@ export * from './LoadingCard';
 export * from './Modal';
 export * from './Pagination';
 export * from './PasswordInput';
+export * from './Popover';
 export * from './Table';
+export * from './Tabs';
 export * from './TypeaheadSelect';